add codeowners support for gerrit
when a Pr object is created from a gerrit change, the sourceBranch in
the Pr points to the branch name renovate uses internally. In the case
of gerrit, this isn't a branch at all (just a a message in the gerrit
change).
During code-owners analysis, the code tries to use the sourceBranch of
the Pr to check what has changed with getBranchFiles.
Since the sourceBranch does not exist, this fails and yields no results.
As a fix, this adds a new property 'commitSha' to Pr.
code-owners checks if commitSha is set, and if so, uses
getBranchFilesForCommit instead, to directly check the diff between this
commit and its parent.